move in memory data stores impl and tests to a specific folder
