+++ backend/src/usecases/add-item-on-inventory/add-item-on-inventory.test.ts
@@ -1,7 +1,7 @@
 import { AddItemOnInventory } from '..'
 import { InventoryItem } from '../../domain/entities'
-import { InMemoryInventoryStore } from '../../data/in-memory-inventory-store'
-import { InMemoryProductStore } from '../../data/in-memory-product-store/in-memory-product-store'
+import { InMemoryInventoryStore } from '../../data/in-memory/in-memory-inventory-store'
+import { InMemoryProductStore } from '../../data/in-memory/in-memory-product-store/in-memory-product-store'
 import { Product } from '../../domain/entities'
 
 describe('Add items on Inventory tests', () => {
